+++ Game/Scene/EnemyGuide/RollBackGround.cpp
@@ -2,7 +2,7 @@
 
 const double RollBackGround::W = 1280;
 const double RollBackGround::H = 720;
-const double RollBackGround::ROLL_SPEED = 5;
+const int RollBackGround::ROLL_SPEED = 5;
 
 
 RollBackGround::RollBackGround(String firstAssetName, String secondAssetName)
@@ -24,13 +24,13 @@ void RollBackGround::update()
 	fPoint_m.x -= ROLL_SPEED;
 	sPoint_m.x -= ROLL_SPEED;
 
-	if (fPoint_m.x <= -0.5*W)
+	if (fPoint_m.x <= static_cast<int>(-0.5*W))
 	{
-		fPoint_m.x = 1.5*W;
+		fPoint_m.x = static_cast<int>(1.5*W);
 	}
-	if (sPoint_m.x <= -0.5*W)
+	if (sPoint_m.x <= static_cast<int>(-0.5*W))
 	{
-		sPoint_m.x = 1.5*W;
+		sPoint_m.x = static_cast<int>(1.5*W);
 	}
 }
